requirements.txt
absl-py==2.1.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.2.4
certifi==2025.4.26
charset-normalizer==3.4.2
flatbuffers==25.2.10
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==2.10.0
idna==3.10
importlib-metadata==6.7.0
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
mock==5.2.0
numpy==1.18.5
oauthlib==3.2.2
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.15.5
tensorflow-estimator==1.15.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
